requirements.txt
# Core bot dependencies
aiogram==3.22.0
mistralai==1.9.10
pyairtable==3.2.0
python-dotenv==1.1.1

# HTTP and async support
aiohttp==3.12.15
httpx==0.28.1
requests==2.32.5

# Data handling
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0

# Type hints and annotations
typing_extensions==4.15.0
annotated-types==0.7.0

# Utilities
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
urllib3==2.5.0
yarl==1.20.1
multidict==6.6.4
frozenlist==1.7.0
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
h11==0.16.0
httpcore==1.0.9
sniffio==1.3.1
anyio==4.10.0
packaging==25.0
six==1.17.0
jiter==0.10.0

# File handling
aiofiles==24.1.0
